pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gwas-mcp"
version = "1.0.0"
description = "A powerful MCP Server for GWAS and Bioinformatics Research - Access 30+ tools from 14 major biological databases"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ZaEyAsa"}
]
keywords = [
    "gwas",
    "bioinformatics",
    "mcp",
    "genomics",
    "snp",
    "variant",
    "protein",
    "uniprot",
    "ensembl",
    "clinvar",
    "alphafold",
    "kegg",
    "string",
    "open-targets",
    "pharmgkb",
    "omim",
    "pdb",
    "claude",
    "llm",
    "ai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "biopython>=1.80",
    "pysam>=0.20.0",
    "pandas-plink>=2.2.0",
    "statsmodels>=0.14.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/zaeyasa/gwas-mcp"
Documentation = "https://github.com/zaeyasa/gwas-mcp#readme"
Repository = "https://github.com/zaeyasa/gwas-mcp"
Issues = "https://github.com/zaeyasa/gwas-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
gwas-mcp = "gwas_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gwas_mcp = ["*.json", "*.yaml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
